+++ CollectionsWithSwiftUI/CollectionsWithSwiftUI/Item.swift
@@ -1,0 +1,14 @@
+//
+//  Item.swift
+//  CollectionsWithSwiftUI
+//
+//  Created by Francisco Soares Neto on 18/11/20.
+//
+
+import SwiftUI
+
+struct Item: Identifiable {
+    let id: String
+    let name: String
+    let imageName: Image
+}
